Big reorganization of %c-step.c, no META/THE/VAR
This does a reorganization of the Meta_Stepper_Executor() to be more
in line with the current "divide things into sections with goto labels"
aesthetic.

It also makes it so that it uses tests of Is_Pinned()/Tied()/Lifted()
rather than assume there are specific Heart_Of() values corresponding
to those states.  This is in preparation for all values having these
states available, and so if you ask for the Heart_Of() you will not
find out about those decorations.