Grand Unification of BLANK and SPACE
We've been pushing slowly but surely into a world where BLANK! was just
like any other part in the system, replacing its "nothingness"
functions" with specialized parts that do nothingness better.

What remained was its need to be nothingness you could put in an array,
but a non-WORD!... reserved so that it could not be assigned over and
serve in places like multi-return blocks or paths/tuples, as "reified
unreassignable nothingness"

In basically all the cases I can think of, something that represents
the literal space character serves just as well as anything for that
kind of nothingness!

This resolves much of the tension over the behavior of blanks as
spaces in dialects, since it is now *the* literal representation for
a space character.