Mark divergent functions with tag
There was a concept that `return: []` meant divergent functions, but
that isn't as useful as being a way of saying the function has no
meaningful return value.  Before this was `return: [~]` but that will
now mean literally a quasar and not an antiform.  So the most sensible
path is just to take the W and make `return: []` the best way of saying
you don't return anything meaningful.