Disallow typechecks on NULL/TRASH by default
The purpose of NULL is to be word-fetchable without an error, but
to basically trigger errors when passed as an argument to anything
that's not supposed to be "conditional".  e.g. the condition of an
IF accepts NULL, but most other arguments should not.

This also makes it so that trash is unfriendly to things like INTEGER?
as well.