Cycle prompt on no input to console
We are now displaying ghosts:

    >> comment "hi"
    == ~,~  ; anti (ghost)

This means if we didn't suppress evaluation of empty transcodes, just
hitting enter in the console would print output:

    >>
    == ~,~  ; anti (ghost)

That is irritating, so this changes it to notice empty transcodes and
cycle the prompt...skipping the evaluation entirely.

To make this easier, empty transcodes return null.