Start reducing Cell* usage
If Cell is to be kept as a separate type, it should only be used in
places where the bit patterns are not known to be initialized to
anything (e.g. not known to even be an Atom*).  That might be better
done with Sink(...), so the cell class could go away...with any
function truly taking any readable data taking Atom*.