Restrict checks on Atom, kill of stable_XXX macros
This moves toward a greater amount of caution on letting functions do
tests on Atoms, which might contain ERROR! values or packs that have
not been decayed, hence giving misleading answers to things (e.g.
Is_Integer() might say a pack with an integer in the first slot is
not an integer, even though it would decay to one).