More strictness: Cell -> Element/Atom/Value
It's not entirely clear what the functional difference between Cell
and Atom should be.  At one time it was the concept of a Cell that
abstractly has no quoting level, so you can't ask it for its type...
only its heart.  That was applied to molding, but this is changing
completely--so it probably will be eliminated.

This gets rid of a lot of the instances, as well as finding places
where binding was being asked of antiforms.