Reorganize scanner to clean up sequence scans
There was some irregularity in sequence scanning.  To get to the bottom
of it this just goes through and cleans up the Scanner_Executor(),
breaking it into labeled sections the way that has become the norm
in Ren-C code.

The major change is to push sigils and quotes on sequence elements
as-is, and then undo the decorations on the head when popping the
sequence from the stack... reapplying them to the popped element.
This solves problems that were occurring from trying to never push
an invalid element.

Also fixes cases related to single character quasiforms.