Make PICK* speak in a meta-protocol
This paves way for the idea that a PICK can give back an ACTION!
(distinct from an ACTION! value that is meant to be the result of
the PICK) in order to handle sub-picking.

The poster child problem is from modeling a C STRUCT!:

    struct.million_ints_field.10

You don't want to expand the int[1000000] packed representation into
a BLOCK! of 1,000,000 INTEGER!, only to pick the 10th one.  The best
way to solve this problem is using a lazy accessor, so then the
(struct.million_ints_field) should return an ACTION! that can respond
to the specific choice to pick the 10th item, or (if not picked)
expand out into whatever representation that is.

This achieves the distinction between returning an ACTION! to act
as a lazy accessor and an ACTION! that is actually the result of the
pick (e.g. field of an object) by using a meta protocol.