Tightens Undecayable Semantics
This enforces the Undecayable rule more often.  It's probably not
the right rule, but gives a chance to explore ideas.

It also brings back the idea that [_]: as a placeholder in multi
returns decays, such that [{_}]: would mandate an ordinary decay
as if the variable had been there... and this applies whether it is
circled or not.  Hence, [^]: is a way to say that it's legitimate
to be undecayable, yet now with the behavior of meta-variables it
passes through the undecayed value non-lifted.