Rework GET/SET with unified PICK*/POKE* (WIP)
Processing TUPLE! in a GET and SET operation were very similar, and
had many of the same concerns.  This takes a step toward a "grand
unification" by carefully managing the cells in play with tailored
calls to PICK* and POKE* running across the forward and writeback
chain of steps.

It's VERY SLOW at the moment, but hopefully will be able to be
optimized more as it becomes the common path for all GET and SET
of TUPLE! and WORD!.  Several new behaviors are added, such as
being able to poke quoted items (map.'(foo): 10) and to use ^META
fields such as (frame.^field)