Cull primordial ACCESSOR experiment
Before the idea of Dual reprsentation in PICKPOKE, the concept of
CELL_FLAG_VAR_IS_ACCESSOR was tried.

However, building the entirety of variable access on top of the
PICK* and POKE* fundamentals.  This only takes away 2 quote states
instead of a whole cell bit, and is much more general.