Corrupt cells at callsite vs. weird name warnings
In an attempt to make it clearer that certain functions were using
level cells as workspaces, their names got long to tell you that.
This shifts the idea from long names to just corrupting the cells
at the callsite, and checking they're corrupt on entry to the
function that does the work.

As this is only in debug builds, it doesn't slow down the release
build behaviors.