Don't try to abstract Dual arg convention
An attempt to make it "easier" to handle dual args arose from trying
to handle unstable void arguments to POKE*.  But it really just meant
paying more cost and getting in the way of optimizations, while not
making things that much clearer.  This takes a different approach to
the problem, trusting people to know how to test for liftedness and
do an unlift.