Switch to consolidated TWEAK merge of PICK*/POKE*
PICK and POKE are often quite similar in terms of finding the address
of something and either modifying it or just returning it.  This
was handled by a unified PD_XXX() function in historical Rebol, which
lacked the GC rigor of generalized function dispatch.  Ren-C has
flipped back and forth on whether this should be some unified mechanism
or not, but had settled on not-unified with PICK* and POKE* low level
functions.

This re-merges the low level routines into a single primitive called
TWEAK* which actually uses dual representation to be able to ask for
things like PROTECT/UNPROTECT as well, or to set fields to act as
accessor functions.