Eliminate Get_Var_May_Panic()
System-wide, "abrupt panics" are being pared down to that which would
not occur in ordinary operation (e.g. out-of-memory errors), and for
things that are reasonable to expect (such as unset variables or an
error during an evaluation) they are bubbled up as Option(Error*).
This makes it possible to use the system on platforms which have no
longjmp/setjmp or C++ exceptions.