Change Init_Trash() => Init_Tripwire()
Calling the antiform of SPACE a "trash" is no longer accurate, as trash
is the name of the entire category.

Hence antiform space is now known as TRIPWIRE.

(There's some question over if lifted tripwires should be called "wire"
or if they should keep the current name of "quasar".  For now,
quasar is kept.)