Force Get of WORD! through 2-step process (slow!)
This is a work in progress, which tries to run more code through a
common path.

The idea is that "steps" on the stack are used to GET a WORD! by using
the binding as the first step.  Unabstractable binding is the new
terra firma for the picking process, implemented through TWEAK*.