Transition APPLY to use SET-XXX not GET-XXX
Using SET-WORD! in dialects was complicated historically by the auto
gathering of "FUNCT" which made variables for anything in a function's
body that was a SET-WORD!.

That bias is gone today, but there was some question over whether the
conflation was desirable inside APPLY.  It seemed to "look nicer" with
slashes.

However, Ren-C evolved a different meaning for slashes, so that instead
of choosing between `/DUP` and `DUP:` the choice was between `:DUP`
and `DUP:`.  That made the SET-WORD more obviously the prettier choice.
But also, questions have arisen on what :XXX will do in the evaluator,
possibly meaning (TRY XXX), in which case it would be far more likely
to be used inside an APPLY at the top level of the dialect than another
SET-WORD.

So this finally buries the hatchet for this long-debated question.  It
also factors out the work done by APPLY so that it is shared by
SPECIALIZE.