Solidify bootstrap typespecs
The previous process of bootstrap was willing to overlook bound but
not yet set tests.  Due to the evaluator's new dependence on TWEAK*
to do word lookups, it's necessary to make bootstrap versions of
TWEAK and NATIVE that do no type checking.  These can then be replaced
with versions that do type checking (though natives only type check
in debug builds).

While doing this, a fuller accounting of out-of-order typespec functions
was made... putting them first in the boot, before the natives that
used them.