Get rid of non-"gotten" uses of Lookup_Word()
The goal is to force all word fetches through common code, that can be
sure to honor things like accessors/typechecking/aliases/etc.  (The
representation will also be lifted in objects, a big change.)

Most difficult to excise (and most punishing performance-wise) will be
the evaluator's optimization of using a "gotten" to check for infix,
and retain that result if it's going to incidentally be the next
thing that gets fetched.

This is a few cases that aren't that.