Fetch full cells for "gotten"
This removes the last usage of Lookup_Word() which didn't use the
GET and SET code paths (that would permit type checking, accessors,
etc.)  It makes it so that the "Gotten" cell in the evaluator which
tries to carry over values between infix lookahead and the next step
is actually a full cell instead of a pointer to some known address.
This means it can cover things like synthesized values that do not
have an actual address (e.g. coming from a GETTER function).

It's slower, and in actuality the asserts may fail once getters are
implemented (because they check the same value is returned from
each call).