Add new Slot type for OBJECT! fields
This creates a new Cell subclass specifically for dealing with fields
of contexts like OBJECT!/MODULE!/LET!.

The idea is to make it incompatible with most functions for doing blind
bit pattern reads or writes on Cell (in order to make allowances for
GETTER and SETTER functions, which may invoke the evaluator).  But
there is one exception, which is that freshly created Slots can be
an Init(Slot) that can be passed to any routine accepting Init(Value)
or Init(Element).