"Lift the Universe" first boot (WIP)
The original formulation of unstable antiforms was that they could
not be stored in variables, unless you lifted them first.

To make handling a bit easier, (^var: ...) was implemented to lift
on assignment, and (^var) would unlift on fetch.

As the conception of how contexts stored variables evolved, it seemed
that it would be better to say that all variables store their "normal"
states lifted, hence unstable states could be stored, but only using
the ^var operators.