Add "true unset" (WIP), make separate WEIRD flag
It seems that the contention between OUT_HINT_UNSURPRISING and
SLOT_HINT_DUAL is enough of a problem to warrant using a distinct
flag.  So this makes CELL_FLAG_WEIRD, to go along with HINT and NOTE.

A fledgling version of dual unset is added, but needs more work before
it will be viable as a replacement for unspecialization.