Stepper_Executor() doesn't lift, no L error flag
With the existence of dual representation and CELL_FLAG_WEIRD, the
it no longer makes sense for the Stepper_Executor() speak in lifted
values instead of ordinary Atom/Slots which--if the Stepper reaches
the end--sets the output to be a tripwire with CELL_FLAG_WEIRD, thus
making it compatible with what a FRAME! would want the state to be.

Also, the LEVEL_FLAG_ERROR_RESULT_OK flag now does more harm than
good.  All code paths need to run through decay anyway, and that's
going to be handled in a smoother fashion by adding Trap_XXX to the
decays that aren't known stable.  It was causing a problem and it
was easier to get rid of it than to figure out what the problem was.