Don't hide Value* in FileReq
This is a crude preservation of behavior in code that mostly just needs
to be rethought completely from an archaic PORT! system.  It used to
store a pointer directly into the port spec object as part of a
FileReq... trusting that the spec would be valid for the duration of
the BLOB! of binary data representing the OS state of the file.  Now
that you don't use direct pointers into objects, it's better to just
regenerate the path from the spec whenever it's needed.

Again... very crude, just trying to keep things working so that more
important things can be attended to.