const references for Option() construct, p rename
The OptionWrapper struct had its member variable named `wrapped` which
is long and annoying to type in the debugger, and doesn't line up with
what all the other pointer wrapping classes call `p`.  Although it
can store integers or enums as well, just call the wrapped field p
for sanity.

Also: it was doing construction by value when it did not need to,
which forced copies at callsites of things like Sink() instead of
being able to use cast operators.  Change that to const reference.