Split NeedWrapper from SinkWrapper
When initially introduced, NeedWrapper and SinkWrapper were similar
enough that it seemed controlling the "corruption" through a flag
template parameter would be a good idea to simplify it.  However,
in practice there are problems with partial specialization of
constructors when you're using something with value semantics of
this kind... and it created an inevitable trap where the copy
constructor was not able to select a direct transformation between
things like Sink(Value) and Need(Atom*)... forcing a temporary copy
of a Sink(Value) to a Sink(Value) introducing a spurious need to
make a corruption pending.

Refactoring the code into a separate SinkWrapper and NeedWrapper,
based on a common PointerWrapper, does not require partially
specialized constructors and can avoid that temporary.