Use struct vs. function specialization for corrupt
The concept behind Corrupt_If_Debug() was that there was a generalized
catch-all corruption which scrambled memory with memset(), and then
there were specializations which handled corrupting things that either
wanted a more performant kind of scrambling, or needed to do something
different for correct semantics.

This faced limitations because Corrupt_If_Debug() was a templated
function, and templated functions can't be partially specialized.

The workaround for this is to use a struct instead, which has a static
function in it...then partially specialize the struct.  This does
that and it means that the partial specializations can carve out
exceptions using SFINAE (e.g. all Cell subclasses can get their
custom corruption in one swoop).