Begin breaking c-enhanced.h up into needful lib
c-enhanced.h became a very large endeavor, big enough that the cast
mechanics were split to their own file.  This breaks it down into more
files, with the idea of factoring it into a library with its own
documentation and tests that could be used by other projects.

It's a partial step toward making that library, while it's decided
exactly what all will go in it.