Suppress error from last-result assign in console
The console has code:

    ^last-result: ^v

In the current formulation, this allows the ^META-parameter V to give
back an ERROR! antiform, which is unstable.  It is then assigned to
the last-result and propagated as ERROR!.

You need to TRY or otherwise triage the result to prevent it from
escalating to a panic.

Also lets you PROBE() an OBJECT! which is storing antiform errors.

    try ^last-result: ^v