cast() checks for Element/Value/Atom, casts dir
This is a long-time-coming change which adds to the cast checking that
cast() of Cell subclasses give valid data for that class (e.g. no
antiforms in Element, no unstable antiforms in Value, and Atom and
Value and Element are all readable at the time of cast...if not use
a u_cast())

An exemption in the casting is required when casting from raw bytes,
because that's what's used to give you potentially tail cells or
uninitialized cells.  But with that exemption, most of the code was
already written correctly for the rules.

This breaks the sys-debug-casts.hpp file into 4 files in a cast
directory and documents it with a README.md