Allow stable antiforms as map keys/values
There wasn't any particular value to prohibiting antiforms in maps,
and it made UPARSE awkward in creating a combinator maps because of
the inability to directly represent datatypes as keys.

This hacks in the ability to use antiforms, and adds a buggy version
of MAKE MAP! that UPARSE can use (needs to check for duplicates, and
either prohibit them or let the last one win).  Moving ahead for now.