Resurrect DEBUG_UTF8_EVERYWHERE
When CI is working again, it really should have a build that tests
the DEBUG_UTF8_EVERYWHERE.

But, for now, this just gets it working again.  Since c_cast() is now
used with Utf8(*) it had to have specializations for the cast helper,
and that cast helper needed to remove the reference type to match
things correctly.

GCC also considers void* casts to be ambiguous, and const char* casts
can be ambiguous with char* casts.  It seems GCC is on the wrong side
of the standard, for once.  Only way to be compatible is to restrict
the implicit casting and add cast operators at callsites that take
void pointers.