New format for %make.r file lists and options
The fact that bootstrap has to load foo.c as foo/c can be worked around
to produce filenames.  So this does that, and also switches from
a format which put files in blocks with their options to having a
block of options after the file.

It's messy--but is a first cut at using PARSE3 to do it.

Note `<opt>` parameters don't work with APPLY in bootstrap.  Review.