Use new literal quasi-group param spec in LOCALE
Things like ~word!~ are too valuable for denoting a desire to match
QUASI-WORD! by example.  So that's the way that is going, vs antiform
of the word "WORD!".  This means also that quoted things like 'word!
are quoted.  So ~(lit-thing)~ is the new idea for literal things.

Adjust LOCALE to this idea.