Bugfix for name caching, hard-line set-only
There was a bug where the name assignment was testing for actions, not
lifted actions (which was what the TWEAK* protocol uses).  So no name
caches were being set on assignments, resulting in a lot of hard to
debug frames.

Finally investigated and fixed it.  And went ahead and made a hard
commitment that the only names used in error reporting are the names
fetched from the action--regardless of the how it is fetched.  This
means a meta assignment that then gets dispatched will have an error
based on the original name.

Since this commitment was made, it simplifies the Begin_Action() code,
which can now be made part of Push_Action().  This simplifies things
a fair bit.

Also makes it so that natives go through normal evaluation for their
assignments.