AUGMENT slots are true unset, not trash
The way AUGMENT was pushing things to the stack it was losing the
CELL_FLAG_SLOT_WEIRD_DUAL, so instead of being true unset the slots
were just trash.

This fixes that and adds a little bit of hackiness to enumeration to
bump along the road toward handling true unset better.