Auto-build TO-XXX functions, no garbage antitypes
The TO-XXX functions used to have to be named explicitly.  Now the LIB
context can be extended manually (and relatively efficiently) with
EXTEND.

This fixes broken creation of shorthands like TO-FILE, and gets rid
of a bunch of dummy antiforms that were being created in the
system.contexts.datatypes due to slots that don't correspond to types
that allow antiforms.