Pass VOID for `<opt>` parameters, not null
In theory an `<opt>` function is not supposed to let you pass a null,
and only accept VOID... which it then receives as NULL for easier
handling.

However, that wasn't being enforced.  A subsequent commit does start
enforcing the rule.

(It's up in the air if void will be a value, and not a function, thus
accessed via ^void instead of void)