Erase_ARG() to permit ARG() to validate
ARG() previously did no validation on the Stable-ness or readability
of the argument, thus allowing it to be garbage and initialized.

But it is pending a need to be validated.  So since the only thing
anyone does with an invalid arg is erase it anyway, substitute the
pattern:

    Erase_Cell(ARG(NAME))

With a new accessor:

    Erase_ARG(NAME)

It's shorter and solves the problem of not validating ARG().