Clear lingering debug protect out on abrupt panic
It's useful to be able to protect the OUT cell to make sure it doesn't
get changed, but an abrupt failure means the cleanup code to remove
the protection never happens.

It may be this bit should be automatically cleared in more places,
but for now just do it on abrupt failures.