Massive simplification: Atom* frame arguments
With the dawn of "Lift the Universe", a new idea of being able to store
unstable antiforms in what act like "normal" values came about.

This takes it to its logical conclusion, where OBJECT!/FRAME!/etc.
actually store Atom* as unstable antiforms as-is.  This means natives
do not have to deal with juggling the lifted state form ^META
arguments, making them as convenient as usermode functions (more or
less) in terms of being able to deal with all variables "as is".

Central to this is a new ARG accessor, Atom_ARG().