Deal with LOCAL() being "pure unset"
The LOCAL() macro had been assuming Value* by default.  That was
leading to asserts, bceause LOCAL() are initialized with a "WEIRD"
dual state of unsetness.

This makes LOCAL() not do any interpretation on the Atom*, and does
some reorganization of functions that were being dodgy about the
initialization state, so that the assigned pointers are never lying.