Fix MSVC optimized build: v_cast() for enum/int
There was a problem with the decay/forward mechanics for v_cast() in
MSVC, where refrences to fundamentals e.g. int& were not matching a
v_cast_decay_helper() instance.

This hacks in a third variant of v_cast_decay_helper() to work around
the problem.  AI didn't seem to have better ideas (making it a struct
helper and using partial specialization seemed to not work due to
running afoul of losing "universal reference" behavior that rvalue
references have with functions).