HEART_BYTE() => KIND_BYTE()
HeartByte doesn't just mean "a Heart cast to a Byte", it now means
specifically holding a Heart multiplexed with a Sigil.

Rather than calling it HeartSigilByte, this makes up a new name...
the KindByte.  This helps as a reminder that it encodes more than
the Heart, kind of like how LiftByte cues you to know it holds more
than just a quoting integer level.