Fix GC bug during type checking
Type checking now can call arbitrary functions, which introduced a
wrinkle that hadn't been accounted for in the GC... this is that if
a frame is in the type checking phase, it should not abbreviate how
far in the frame it marks based on the enumeration position of the
frame's key/param/arg.

The issue is that there's a single flag controlling whether or not
an action frame is dispatching...and once it is, the LEVEL_STATE_BYTE
belongs to the native/dispatcher.  So you can only ascribe meaning
to the state byte when not in dispatch.  The dichotomy that had been
set up was Is_Level_Dispatching() vs. Is_Level_Fulfilling() -BUT-
it was really all non-dispatch states (including type checking) that
were covered by the non-dispatch.  So more accurately it would be
Is_Level_Fulfilling_Or_Typechecking().

This needs review, but the situation was clearly wrong and causing
hard-to-find bugs because the GC wasn't marking all the slots when
the type checking was in the early part of the check.  So the major
problem is fixed here.