Tune x_cast() and reassign to m_cast()
The runtime costs in the debug build of some of the Needful abstractions
are small, but enough to cost a lot when using them with a large number
of low-level operations.  So if something like a cast operator can be
done without a function template (which has cost in debug builds even
if it is zero cost in release builds) it's best to do it without the
function template.

It turns out that a relatively safe expanded form of mutability cast
can be done without a function template.  It won't work for targeting
a pointer wrapper class, but it can work most of the time.  This
rethinks x_cast() as that safe version and gives it the name m_cast(),
which can be cheap-as-free, even in debug builds.

Then w_cast() is for the "writable wrapper" conversion, which only
needs to be used occasionally.