Make ensure() cost nothing in debug builds
ensure() is for use in macros when you want to do type checking
without having to make an INLINE function helper to do it.

However, it was implemented via function templates.  This meant that
in debug builds, it had the same cost as an inline function--just a
bit more convenient in a macro.

BUT this commit changes it so that ensure() actually has -no- function
template, and thus can do its check with compile-time cost only in
the debug build, vs. just the release build.

It also makes it more pleasant to step into code that has ensure(),
because there's no debug step for it.