Cast hook temporary disable/enable by #include
This adds a feature by which you can disable cast hooks, so in the span
you disable them you don't have to use the u_xxx() variations of the
cast... it can simply be understood that in a certain file range you
don't want the hooks because that code is performance sensitive.

This means you can save the unhooked casts for cases that are
semantically non-checkable, e.g. the bits are not well-formed.  You
can easily turn the checks back on if you want to do a "heavy" check
of the performance-sensitive code, without touching all the cast sites.