Push all LIFT_BYTE() values up by one
This changes it to where LIFT_BYTE() has a new baseline of DUAL_0.

The plan is to use this to get rid of CELL_FLAG_WIERD_DUAL, since the
dual state would only take up one value in the lift byte, and this
makes slots more compatible generally with the rest of the code.

The actual change to get rid of CELL_FLAG_WEIRD_DUAL is a bit involved
so it's a separate commit.