Transition to *UNSET* for dual unset signal
Previously it was an antiform (tripwire).  But choosing something that
is not an antiform, not quasi, and not quoted allows the representation
among lifted items in a block, without conflation:

    [a b c]: ~[~null~ *unset* '1020]~

The asterisks seem like about the best way to make it stand out, while
still having the quickness of testing like with a symbol.