First cut at replacing SLOT_WEIRD_DUAL w/DUAL_0
SLOT_WEIRD_DUAL took up a whole flag, when really it just needed to
chew out one state of the lift byte...that makes everything more
coherent.

This takes the first step of replacing the flag, but with the problem
that Atom* should not be holding DUAL_0 LIFT_BYTE()s.  (This may need
to be an exception.)