Speed up isotopic conversions to quasi/anti forms
In the general case, we don't allow arbitrary Cells with arbitrary
hearts/sigils to be turned into antiforms or quasiforms.  So UNLIFT
and QUASI have to go through a special coercion... it removes bindings
from antiforms, and ensures that the heart is compatible and that
there is no sigil for an antiform.

But this has cost.  If we know we are doing a legal conversion with
a known heart byte and no Sigil or binding, then it should be free.

To make it free and safe, this adds routines with checks to make
sure you know what you're doing:

   Stably_Antiformize_Unbound_Fundamental()
   Unstably_Antiformize_Unbound_Fundamental()