Use more caution testing Atom* for stable states
Helpers like Is_Atom_Action() or Is_Atom_Trash() were added without
siutably pointing out how questionable it is to test possibly unstable
antiform states for stable states.  e.g. it could be a PACK! and
would decay to the thing you were testing for.  Or it could be an
ERROR! and you shouldn't be able to test it at all.

This pares back the usage, but doesn't eliminate it, as it has a
slight edge of efficiency.