Cast fixes for non-debug GCC build, kill d_cast()
Doing a build for Valgrind put GCC in a mode where it complained about
some casts, so this fixes that.  It also gets rid of the d_cast()
idea... Needful is not for code that "thinks like C++" so having a
cast that's explicitly for a cast hierarchy was a mistake... it just
adds complexity and any real value is given by the hooking in casts
that will tell you when the cast is valid or not.  It's diminishing
returns to document places that are downcasts: if you didn't think
the cast was valid, you wouldn't be doing it, so just let the hooks
handle checking that.