Fix scanner uninitialized use found by valgrind
Valgrind finds some bugs address sanitizer doesn't--this was one.

Before the TranscodeState and ScannerState had been clearly split up,
it was hard to understand what the variables meant.  But this has
an incorrect usage of the ScannerState token ->begin, which is only
valid when a token is being scanned.  What was wanted was the current
transcode position shared across all scannerstates for that transcode.