Fix surprising action hijack bug
Valgrind noticed that there was an access of the optimized type byte
out of range (if there were no type check bytes, you can't check the
seocnd type range byte for 0).

Fixing that led to another realization, that if there were unoptimized
types to account for it should make the result "surprising".  This
meant HIJACK had to be changed to not return tripwire if you hijack
something that's been previously hijacked as void, but rather you
have to return some function.  This returns a new UNIMPLEMENTED
native--but the code should be reviewed to see if it can be more
elegantly done.