Reorganize main() after valgrind wild goose chase
Valgrind had some bad reporting of perceived leaks, which turned out
to be system libraries that had reachable pointers.

Because the reporting didn't have parity with Address Sanitizer in
a debug configuration, this led to researching why...and it turns out
that Address Sanitizer has a feature kind-of like Valgrinds, but it
doesn't have any way to filter the report.

Rather than throw away the work, this documents it, and makes the
organization of main.c somewhat better in terms of phases.