Rename String* Flex type to Strand*
This avoids confusion with the userspace notion of ANY-STRING? for the
internal type that's a Stub (not a Cell) tracking the memory behind
the strings.

It's a good name because it can still work with str or s variables,
and evokes something "string adjacent", while not being a full-on
String in and of itself (no index, or other Cell properties).