Use cell bit to test for null faster
This improves conditional testing by making it so that if a KEYWORD!
antiform WORD! cell is initialized with the symbol for NULL, it also
puts a bit in the header saying it's null.  This makes testing for
null only require looking at the header... becoming faster across
the board in IF statements or ELSE or anywhere else that looks at null.