Simplify %needful-corrupt.h
There were ideas of being able to corrupt a pointer and then check if
it was a corrupt value, which is fundamentally at odds with the idea
of being able to occasionally corrupt to a zeroed pointer.  There
was also complication to try and make it possible for the C99
corruption to be more efficient than a memset(), which ran afoul of
type assignments to pointer etc. and needed workarounds that weren't
worth it.

Also, there was Corrupt_Pointer_If_Debug() as a syntax in case you
didn't realize that Corrupt_If_Debug() was going to corrupt the
pointer itself vs. the pointed to thing... but this added complexity
for no particularly good reason when people need to know what it
does anyway.

Name also changed to Corrupt_If_Needful, which is funny and sounds
like Corrupt_If_Needed, and doesn't entangle with whatever flags
you are using to indicate debug builds / checked builds / etc.