libuv => 513751e2fcf1b5be5238de66b2c06f3e0623aca0
This does an update of libuv to whatever state it is in now in
June 2025.

(Motivation was that old version did not support memory sanitizer,
e.g. it would make syscall()s and did not unpoison the structures
so as far as MSAN was concerned the memory wasn't initialized,
even though the OS had written information into them.)