Merge `fail` Result with `FAIL` Bounce
The innovation of the Result type is able to be polymorphic such that
it can work with Bounce as well.  However, it makes the nullptr state
of Bounce do double duty--if the g_failure error is not set, that
means the user intended a ~null~ antiform.  But if the failure is
set, it was a `return fail` and not a `return nullptr`