Expand conversions from Trap_Xxx to Result(T)
This pushes forward on using the new trap and except features.

It doesn't do the right discernment between panic() and return fail(),
and the main trampoline doesn't know how to respond to zero results.

Folding the functionality for concretizing bounces into the trampoline
seemed like a bad idea at first but it is now seeming like a better
idea.