Add OPAQUE! type for GOB!, `!` names in %types.r
If you are using Cell in the implementation of your extension types,
then you might want to have the opportunity to use nearly all the
bits in the Cell and not pay anything for the type information, for
parts that are not exposed to the user.  This was done in GOB! for
the "XYF", but other types might want to use the ability.