Put all C definitions for Needful in needful.h
This re-styles the needful library in such a way that needful.h is a
full definition of the nearly no-op behaviors of all the constructs
when built as C.

Consequently, it's much easier for people to see what it's doing, and
also to possibly just add one file to their project in order for it
to build.  They can then have their continuous integration or local
build settings enlist separately in the Needful project for the
extra behaviors to build with them.

In addition this makes some advances on having needful not pollute
the namespace by default, and allowing you to opt-out of defining
names that aren't prefixed by needful for all constructs.  The C++
helpers use a C++ namespace so they can define as many things as
they want to factor things for clarity.