Rename Nothing to Zero
Plan to use "none" for disengaged optionals means nothing looks too
close to it.  Other things like "nonce" would be worse.

Zero is a decent enough name.

It's not possible to constrain construction to just the literal 0,
so use `zero`.  Refine methods of defining zero in a non-contentious
way with fields or variables named zero via constexpr/enum.