Drop NeverNull from Needful
The main reason NeverNull was kept around was to serve as a template
for writing debug classes that would track modifications to a
variable.  But not only have I gotten better at doing that because
I "just know how", AI can create these for you--no need to have a
dead class hanging around.

More generally, "never null" is the general assumption in interfaces
for a raw pointer when it's not wrapped in Option(), which is the
natural way to bias it.  So there's not going to be a reversal of
that where all raw pointers are replaced with NeverNull() pointers,
so this really was just a lingering debug tool.