Rename stub "leader" field to just "header"
In order to write fast macros that didn't use inline functions to
verify the right type being passed in, the header in Cell was called
`header` while the one in stub was called `leader`.  The naming oddity
had to be explained repeatedly.

Now that ensure() has been established as a pure compile-time only
tool, it can type check the arguments to macros.  Trusting ensure()
and changing the name to header makes it clearer, and that these
headers are in the same position.  The code looks more sane, and
there's no need to keep explaining what a "leader" is.