Don't implicitly convert option to int
In order to facilitate usages in switch() statements, Option() would
implicitly coerce to integers (possibly 0).

That idea predated the modern nice syntax of `maybe` to unwrap with
tolerance of zero states.  This is the better way to go, and then
you can have Option(int), which correctly does not implicitly
coerce to int.

(This also avoids having to specialize to avoid conflicts with the
coercion operator when you do specialize to int.)