w_cast/m_cast merge, raw pointer compile-time only
This accomplishes something that was hoped to be accomplished, which
is to use the same operator for mutably casting both wrapped types
and non-wrapped... but not bring in cost to the debug build for
the non-wrapped types.  A macro static_casts the expression to its
wrapped type in const form (if it's wrapped), and then const_casts
away the const, and then static_casts back to the target type.

This is dependent on being able to get information out of the wrapped
type statically of what type it wraps, so the wrapped types must
expose a wrapped_type field.