Re-bias ensure() that mutable accepts const
The ensure() construct would let you say ensure(const char*, ...) if
you wanted to permit both const and mutable, but ensure(char*, ...) if
you wanted only mutable.

That makes it act more like a function parameter.  But in practice it
is more useful if when you don't specify const that you are willing
to match and pass thru either the const or the non-const form.

Both versions are added here, but the lenient form gets the default
name of ensure().  But it makes it so that if you say const on the
target type it will be const.

This also changes the merge const semantics for u_cast(), such that
if you `u_cast(const T, ...)` you will get a const type out.  That
makes its behavior consistent with lenient ensure.